backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph==1.0.10
langchain-core==1.2.17
langchain-openai==1.1.10
tavily-python==0.7.22
arxiv==2.4.1
feedparser==6.0.12
beautifulsoup4==4.12.3
requests==2.32.5
pydantic==2.12.5
python-dotenv==1.0.0
sse-starlette==1.8.2
